Feature/generate from version 1.211.0 (#42)
* Skip flaky test

* Update relewise client to 1.211.0

* Generated the new moddels

* Updated the search hightlight test to use the new model

* Single is a float not an int

* Floats instead of ints

* Date time should be handled like date time

* A struct is also a value type so we should not go into this if when type is struct

* Also handle key value pair generically

* new models after hydration changes

* Add a test proving recently purchased facet not working

* Add a non generated time span

* Do not use depricated stuff

* Make custom types more generic

* Optional generic lists cannot have a defualt value

* new recently purches facet model

* rename test

* Override chosen constructor for `FieldIndexConfiguration`.

* Updated unit test to fit old constructor again.

* Review comments

* Remove timespan

* Rename test

---------

+++ src/Models/LineItem.php
@@ -6,11 +6,11 @@ class LineItem
 {
     public Product $product;
     public ?ProductVariant $variant;
-    public int $quantity;
+    public float $quantity;
     public float $lineTotal;
     public ?array $data;
     
-    public static function create(Product $product, ?ProductVariant $variant, int $quantity, float $lineTotal) : LineItem
+    public static function create(Product $product, ?ProductVariant $variant, float $quantity, float $lineTotal) : LineItem
     {
         $result = new LineItem();
         $result->product = $product;
@@ -62,7 +62,7 @@ class LineItem
         return $this;
     }
     
-    function setQuantity(int $quantity)
+    function setQuantity(float $quantity)
     {
         $this->quantity = $quantity;
         return $this;
